backend/pyproject.toml
[project]
name = "feedbackos-agent-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-multipart>=0.0.9",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "python-docx>=1.1",
  "langgraph>=0.2",
  "httpx>=0.27",
  "redis>=5.0",
  "pymilvus>=2.4"
]

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
